feat: refactor map components and introduce custom hooks for search and feature management
